feat!: simplify program architecture by combining Web and Api layers
