+++ ServiceDirectory.sln
@@ -8,8 +8,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceDirectory.Domain", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceDirectory.Presentation.Web", "src\ServiceDirectory.Presentation.Web\ServiceDirectory.Presentation.Web.csproj", "{75BA5491-0754-496E-8B37-3458A1E2CC6A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceDirectory.Presentation.Api", "src\ServiceDirectory.Presentation.Api\ServiceDirectory.Presentation.Api.csproj", "{931BBCFE-B292-40BD-BF73-209B47CAD38E}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{EDF11955-66C3-4E9D-B55B-9F003543EF04}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{DB01415C-126A-4861-A138-BCFEA1AD8D8A}"
@@ -22,7 +20,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceDirectory.Infrastruc
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceDirectory.Application.UnitTest", "test\ServiceDirectory.Application.UnitTest\ServiceDirectory.Application.UnitTest.csproj", "{5DB3A22C-09C9-424C-8F93-458B4AC3EDB7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceDirectory.Presentation.Api.IntegrationTest", "test\ServiceDirectory.Presentation.Api.IntegrationTest\ServiceDirectory.Presentation.Api.IntegrationTest.csproj", "{031CC3AF-7288-415F-A691-9B9D61104E37}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceDirectory.Presentation.Web.IntegrationTest", "test\ServiceDirectory.Presentation.Web.IntegrationTest\ServiceDirectory.Presentation.Web.IntegrationTest.csproj", "{031CC3AF-7288-415F-A691-9B9D61104E37}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -46,10 +44,6 @@ Global
 		{75BA5491-0754-496E-8B37-3458A1E2CC6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{75BA5491-0754-496E-8B37-3458A1E2CC6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{75BA5491-0754-496E-8B37-3458A1E2CC6A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{931BBCFE-B292-40BD-BF73-209B47CAD38E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{931BBCFE-B292-40BD-BF73-209B47CAD38E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{931BBCFE-B292-40BD-BF73-209B47CAD38E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{931BBCFE-B292-40BD-BF73-209B47CAD38E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7A23B8C0-8900-4295-8417-8A56ECA88BAC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7A23B8C0-8900-4295-8417-8A56ECA88BAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7A23B8C0-8900-4295-8417-8A56ECA88BAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -75,7 +69,6 @@ Global
 		{C265A864-8288-4B99-BC8A-981DE8884C4A} = {EDF11955-66C3-4E9D-B55B-9F003543EF04}
 		{F2FC1C5C-1BD1-470B-B646-60546EEE5237} = {EDF11955-66C3-4E9D-B55B-9F003543EF04}
 		{433BCC4D-ECCF-4901-BC82-45E41CCB0690} = {EDF11955-66C3-4E9D-B55B-9F003543EF04}
-		{931BBCFE-B292-40BD-BF73-209B47CAD38E} = {EDF11955-66C3-4E9D-B55B-9F003543EF04}
 		{75BA5491-0754-496E-8B37-3458A1E2CC6A} = {EDF11955-66C3-4E9D-B55B-9F003543EF04}
 		{7A23B8C0-8900-4295-8417-8A56ECA88BAC} = {DB01415C-126A-4861-A138-BCFEA1AD8D8A}
 		{EA963C2B-A69F-44AB-89C4-6AA13E910BBC} = {EDF11955-66C3-4E9D-B55B-9F003543EF04}
