--- src/ServiceDirectory.Presentation.Api/Endpoints/MinimalPostcodeEndpoints.cs
@@ -1,15 +1,0 @@
-using ServiceDirectory.Application.Postcode.Queries;
-
-namespace ServiceDirectory.Presentation.Api.Endpoints;
-
-public class MinimalPostcodeEndpoints
-{
-    public void Register(WebApplication webApplication)
-    {
-        webApplication.MapGet("/postcode/location/{postcode}",
-            async (string postcode, IPostcodeQuery postcodeQuery) =>
-                await postcodeQuery.GetLocationFromPostcode(postcode))
-            .WithOpenApi()
-            .WithDescription("Get the user's latitude and longitude from a postcode.");
-    }
-}
