--- src/ServiceDirectory.Presentation.Api/ServiceDirectory.Presentation.Api.csproj
@@ -1,26 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net9.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.OpenApi" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Design">
-          <PrivateAssets>all</PrivateAssets>
-          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" />
-        <PackageReference Include="NSwag.AspNetCore" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\ServiceDirectory.Application\ServiceDirectory.Application.csproj" />
-      <ProjectReference Include="..\ServiceDirectory.Infrastructure.Data\ServiceDirectory.Infrastructure.Data.csproj" />
-      <ProjectReference Include="..\ServiceDirectory.Infrastructure.Postcode\ServiceDirectory.Infrastructure.Postcode.csproj" />
-    </ItemGroup>
-
-</Project>
