--- src/ServiceDirectory.Presentation.Api/appsettings.json
@@ -1,12 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "Database": "Server=localhost,1433;Database=ServiceDirectory;User Id=sa;Password=QbLJzwGblAR3USLi;Encrypt=True;TrustServerCertificate=True"
-  }
-}
