--- src/ServiceDirectory.Presentation.Web/Client/IApiClient.cs
@@ -1,8 +1,0 @@
-using ServiceDirectory.Domain.Postcode;
-
-namespace ServiceDirectory.Presentation.Web.Client;
-
-public interface IApiClient
-{
-    public Task<Location?> GetLocationFromPostcode(string postcode);
-}
