+++ src/ServiceDirectory.Presentation.Web/Program.cs
@@ -1,24 +1,66 @@
 using GovUk.Frontend.AspNetCore;
-using ServiceDirectory.Presentation.Web.Client;
+using Microsoft.EntityFrameworkCore;
+using NSwag;
+using ServiceDirectory.Application.Database.Commands;
+using ServiceDirectory.Application.Postcode.Queries;
+using ServiceDirectory.Infrastructure.Data;
+using ServiceDirectory.Infrastructure.Postcode;
+using ServiceDirectory.Presentation.Web.Endpoints;
 
 namespace ServiceDirectory.Presentation.Web;
 
-public static class Program
+public class Program
 {
-    public static void Main(string[] args)
+    public static async Task Main(string[] args)
     {
         WebApplicationBuilder builder = WebApplication.CreateBuilder(args);
 
         builder.Services.AddRazorPages();
         builder.Services.AddGovUkFrontend();
         
-        // TODO: URLs will be stored in appsettings.json
-        builder.Services.AddHttpClient<IApiClient, ApiClient>(client =>
+        builder.Services.AddEndpointsApiExplorer();
+        builder.Services.AddOpenApiDocument(options =>
         {
-            client.BaseAddress = new Uri("https://localhost:7086");
+            options.PostProcess = document =>
+            {
+                document.Info = new OpenApiInfo
+                {
+                    Version = "Version 1.0",
+                    Title = "Service Directory - API",
+                    Description = "The API component of the Service Directory Web Application"
+                };
+            };
+        });
+        
+        // TODO: URLs will be stored in appsettings.json
+        builder.Services.AddHttpClient<IPostcodeClient, PostcodeClient>(client =>
+        {
+            client.BaseAddress = new Uri("https://api.postcodes.io");
+        });
+        
+        builder.Services.AddTransient<IPostcodeQuery, PostcodeQuery>();
+        builder.Services.AddTransient<IMockDataCommand, MockDataCommand>();
+        
+        builder.Services.AddSingleton<MinimalServiceEndpoints>();
+        
+        builder.Services.AddDbContext<IApplicationDbContext, ApplicationDbContext>(options =>
+        {
+            options.UseSqlServer(builder.Configuration.GetConnectionString("Database"));
+            options.UseQueryTrackingBehavior(QueryTrackingBehavior.NoTrackingWithIdentityResolution);
         });
 
         WebApplication app = builder.Build();
+        
+        using (IServiceScope scope = app.Services.CreateScope())
+        {
+            if (builder.Environment.IsDevelopment())
+            {
+                ApplicationDbContext applicationDbContext = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
+                await applicationDbContext.Database.MigrateAsync();
+            }
+            
+            scope.ServiceProvider.GetRequiredService<MinimalServiceEndpoints>().Register(app);
+        }
 
         if (!app.Environment.IsDevelopment())
         {
@@ -26,12 +68,16 @@ public static class Program
             app.UseHsts();
         }
 
+        app.UseOpenApi();
+        app.MapOpenApi();
+        app.UseSwaggerUi();
+        
         app.UseHttpsRedirection();
+        
         app.UseRouting();
-        app.UseAuthorization();
         app.MapStaticAssets();
         app.MapRazorPages().WithStaticAssets();
 
-        app.Run();
+        await app.RunAsync();
     }
 }
