+++ src/ServiceDirectory.Presentation.Web/ServiceDirectory.Presentation.Web.csproj
@@ -8,9 +8,18 @@
 
     <ItemGroup>
       <PackageReference Include="GovUk.Frontend.AspNetCore" />
+      <PackageReference Include="Microsoft.AspNetCore.OpenApi" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore.Design">
+        <PrivateAssets>all</PrivateAssets>
+        <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      </PackageReference>
+      <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" />
+      <PackageReference Include="NSwag.AspNetCore" />
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\ServiceDirectory.Application\ServiceDirectory.Application.csproj" />
       <ProjectReference Include="..\ServiceDirectory.Domain\ServiceDirectory.Domain.csproj" />
     </ItemGroup>
