+++ src/ServiceDirectory.Presentation.Web/appsettings.json
@@ -5,5 +5,8 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "Database": "Server=localhost,1433;Database=ServiceDirectory;User Id=sa;Password=QbLJzwGblAR3USLi;Encrypt=True;TrustServerCertificate=True"
+  }
 }
